Fixing some C++ conversion errors.
* Implicit conversion from int to bool.
* Implicit conversion from size_t to uint32_t.
* Implicit conversion from char* to uint8_t.

Adding no-op color operators so unhandled platforms can still link.

+++ source/text.cpp
@@ -176,7 +176,7 @@ spv_result_t encodeImmediate(libspirv::AssemblyContext* context,
                                "Invalid immediate integer: !"))
     return error;
   context->binaryEncodeU32(parse_result, pInst);
-  context->seekForward(strlen(text));
+  context->seekForward(static_cast<uint32_t>(strlen(text)));
   return SPV_SUCCESS;
 }
